requirements.txt
contextlib2==0.5.1
Flask==0.10.1
funcsigs==0.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mock==1.3.0
pbr==1.8.1
py==1.4.31
pytest==2.9.1
PyYAML==3.11
redis==2.10.5
requests==2.9.1
selenium==2.53.1
six==1.10.0
vcrpy==1.7.4
vcrpy-unittest==0.1.4
Werkzeug==0.11.5
wrapt==1.10.7
